SC 110075: Separate batched components (part 6)
