+++ src/applications/appeals/shared/components/ContestableIssues.jsx
@@ -11,14 +11,11 @@ import { focusElement, scrollTo } from 'platform/utilities/ui';
 import { LAST_ISSUE, MAX_LENGTH, REVIEW_ISSUES, SELECTED } from '../constants';
 import { FETCH_CONTESTABLE_ISSUES_FAILED } from '../actions';
 import { IssueCard } from './IssueCard';
-import {
-  ContestableIssuesLegend,
-  ApiFailureAlert,
-  NoEligibleIssuesAlert,
-  NoneSelectedAlert,
-  MaxSelectionsAlert,
-  removeModalContent,
-} from '../content/contestableIssues';
+import { removeModalContent } from '../content/contestableIssues';
+import { ContestableIssuesLegend } from './ContestableIssuesLegend';
+import { MaxSelectionsAlert } from './MaxSelectionsAlert';
+import { NoneSelectedAlert } from './NoneSelectedAlert';
+import { MessageAlert } from './MessageAlert';
 import { focusIssue } from '../utils/focus';
 import {
   calculateIndexOffset,
@@ -206,8 +203,22 @@ const ContestableIssues = props => {
   return (
     <>
       <div name="eligibleScrollElement" />
-      {showApiFailure && <ApiFailureAlert />}
-      {showNoneSelected && <NoEligibleIssuesAlert />}
+      {showApiFailure && (
+        <MessageAlert
+          title="We can’t load your issues right now"
+          message={`You can come back later, or if you’d like to add your issue manually, you can select "Add a new issue" to get started.`}
+          errorKey="api_load_error"
+          errorReason="API load error"
+        />
+      )}
+      {showNoneSelected && (
+        <MessageAlert
+          title="Sorry, we couldn’t find any eligible issues"
+          message={`If you’d like to add your issue for review, select "Add a new issue" to get started.`}
+          errorKey="no_eligible_issues_loaded"
+          errorReason="No eligible issues loaded"
+        />
+      )}
       {showEditModeError && (
         <NoneSelectedAlert
           count={formData.contestedIssues?.length || 0}
